extract interfaces to the shared folder
