+++ server/services/movies.service.ts
@@ -1,6 +1,6 @@
-import mockMovies from '@app/mockData/movies';
-import { IMovie, Genres, IMovieSortOptions } from '@app/mockData/movies.model';
-import { ModalType } from '@app/components/modals/coreModal/coreModal.interface';
+import mockMovies from '@server/mockData/movies';
+import { ModalType } from '@shared/interfaces/coreModal.interface';
+import { IMovie, Genres, IMovieSortOptions } from '@shared/interfaces/movies.model';
 
 class MovieService {
   movies: Array<IMovie>;
