+++ src/components/modals/deleteMovieModal/deleteMovieModal.component.tsx
@@ -1,7 +1,7 @@
 import React, { FC } from 'react';
 import Button from '@app/components/button/button.component';
+import { IModal } from '@shared/interfaces/coreModal.interface';
 import { ButtonType } from '@app/components/button/button.interface';
-import { IModal } from '@app/components/modals/coreModal/coreModal.interface';
 import useStyle from '@app/components/modals/deleteMovieModal/deleteMovieModal.component.style';
 
 const DeleteModal: FC<IModal> = ({ onCancelClick, onConfirmClick }) => {
