+++ src/components/multipleDropdown/multipleDropdown.interface.ts
@@ -1,4 +1,4 @@
-import { Genres } from '@app/mockData/movies.model';
+import { Genres } from '@shared/interfaces/movies.model';
 
 export default interface IMultipleDropdown {
   genres: Array<Genres>;
