+++ src/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from '@app/components/app/app.component';
-import 'index.css'; // TODO: import styles directly into index.html using webpack
+import 'src/index.css'; // TODO: import styles directly into index.html using webpack
 
 ReactDOM.render(<App />, document.querySelector('#root'));
